Finally added a upload profile picture system.
It adds the filename to the user that is editing their profile and i use the filename to display the image. Still no Cropping though and still doesn't have an upload system for the cover. Because I don't know how to set a certain aspect ration and crop an image.